v4

+++ backend/ingestion/schemas.py
@@ -5,4 +5,4 @@ class AudioPayload(BaseModel):
     device_id: str
     timestamp: float = Field(default_factory=lambda: 0.0) # Will be set to server time if 0.0 or missing
     session_id: str
-    samples: List[int]
+    samples: List[float]
